+++ src/extension.ts
@@ -24,8 +24,7 @@ export function activate(context: vscode.ExtensionContext) {
 
     // Start Target Management UI
     new TargetUiManager().start(context);
-
-
+ 
     // Exclude leaf-data from file watcher
     let config = vscode.workspace.getConfiguration(undefined, null);
     config.update("files.watcherExclude", { "**/leaf-data/**": true }, vscode.ConfigurationTarget.Global);